Move the fake image generation
