+++ tests/test_image_stack_py.py
@@ -2,13 +2,7 @@ import logging
 import numpy as np
 import unittest
 
-from kbmod.core.image_stack_py import (
-    image_stack_add_fake_object,
-    make_fake_image_stack,
-    ImageStackPy,
-    LayeredImagePy,
-)
-from kbmod.core.psf import PSF
+from kbmod.core.image_stack_py import ImageStackPy, LayeredImagePy
 
 
 class test_layered_image_py(unittest.TestCase):
@@ -310,73 +304,6 @@ class test_image_stack_py(unittest.TestCase):
         matched_inds = stack.get_matched_obstimes(query_times, threshold=0.01)
         expected = [-1, 0, 0, -1, 1, 4, 5, 5, 7, 7, -1]
         self.assertTrue(np.array_equal(matched_inds, expected))
-
-    def test_make_fake_image_stack(self):
-        """Test that we can create a fake ImageStackPy."""
-        fake_times = np.arange(10)
-        fake_stack = make_fake_image_stack(200, 300, fake_times)
-        self.assertEqual(fake_stack.num_times, 10)
-        self.assertEqual(fake_stack.width, 300)
-        self.assertEqual(fake_stack.height, 200)
-        self.assertEqual(fake_stack.npixels, 200 * 300)
-        self.assertEqual(fake_stack.total_pixels, 10 * 200 * 300)
-        self.assertTrue(np.all(fake_stack.num_masked_pixels() == 0))
-        self.assertEqual(len(fake_stack.sci), 10)
-        self.assertEqual(len(fake_stack.var), 10)
-        for idx in range(10):
-            self.assertEqual(fake_stack.sci[idx].shape, (200, 300))
-            self.assertEqual(fake_stack.var[idx].shape, (200, 300))
-            self.assertTrue(len(np.unique(fake_stack.sci[idx])) > 1)
-            self.assertTrue(np.allclose(fake_stack.var[idx], 4.0))
-        self.assertEqual(len(fake_stack.psfs), 10)
-
-    def test_image_stack_add_fake_object(self):
-        """Test that we can inset a fake object into an ImageStackPy."""
-        num_times = 5
-        height = 200
-        width = 300
-
-        fake_times = np.arange(num_times)
-        sci = np.full((num_times, height, width), 0.0)
-        var = np.full((num_times, height, width), 1.0)
-        psfs = [PSF.from_gaussian(0.5) for i in range(num_times)]
-        fake_stack = ImageStackPy(fake_times, sci, var, psfs=psfs)
-
-        image_stack_add_fake_object(fake_stack, 50, 60, 1.0, 2.0, 100.0)
-        for t_idx, t_val in enumerate(fake_times):
-            # Check that we receive a signal at the correct location that
-            # is non-zero but less than the flux (due to the PSF) at each time step.
-            px = int(50 + t_val + 0.5)
-            py = int(60 + 2.0 * t_val + 0.5)
-            self.assertGreater(fake_stack.sci[t_idx][py, px], 50.0)
-            self.assertLess(fake_stack.sci[t_idx][py, px], 100.0)
-
-            # Far away from the object, the signal should be zero.
-            self.assertAlmostEqual(fake_stack.sci[t_idx][30, 40], 0.0)
-
-    def test_get_masked_fractions(self):
-        """Test that we can compute the fraction of pixels masked at each index."""
-        num_times = 20
-        height = 20
-        width = 30
-        stack = ImageStackPy()
-
-        # Append images with an increasing fraction of masked pixels
-        for idx in range(num_times):
-            sci = np.full((height, width), float(idx))
-            var = np.full((height, width), 0.1 * float(idx))
-            msk = np.zeros((height, width))
-            msk[0:idx, :] = 1
-            stack.append_image(idx / 20.0, sci, var, mask=msk)
-
-        fractions = stack.get_masked_fractions()
-        self.assertTrue(np.allclose(fractions, np.arange(num_times) / float(num_times)))
-
-        # Filter out images with above a given masked fraction.
-        stack.filter_images(fractions < 0.5)
-        times_remaining = int(num_times / 2)
-        self.assertEqual(stack.num_times, times_remaining)
-        self.assertTrue(np.allclose(stack.times, np.arange(times_remaining) / float(num_times)))
 
     def test_mask_by_science_bounds(self):
         """Test that we can mask pixels by their science values."""
